+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 - All StockMoveLines now appear in Produced products grid (ManufOrder)
 - Fix the default amount on new invoice payment to use the amount remaining of the invoice.
 - Fix demo data en and fr on AppSuplychain to set the correct value on the field supplStockMoveMgtOnSO
+- Fix event hide when we create new from calendar.
+- Calendar display all events linked to the same team that the active team of the user.
 
 ## [Unreleased 4.x]
 ### Improvements